requirements.txt
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==2.1.0
scikit-learn==1.5.1